+++ src/Translatable/Translatable.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Dimsav\Translatable;
+namespace Etchfoda\Translatable;
 
 use Illuminate\Support\Str;
 use Illuminate\Database\Eloquent\Model;