+++ tests/models/CountryWithCustomTranslationModel.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Dimsav\Translatable\Test\Model;
+namespace Etchfoda\Translatable\Test\Model;
 
-use Dimsav\Translatable\Translatable;
+use Etchfoda\Translatable\Translatable;
 
 class CountryWithCustomTranslationModel extends Country
 {
@@ -10,5 +10,5 @@ class CountryWithCustomTranslationModel extends Country
 
     public $table = 'countries';
     public $translationForeignKey = 'country_id';
-    public $translationModel = 'Dimsav\Translatable\Test\Model\CountryTranslation';
+    public $translationModel = 'Etchfoda\Translatable\Test\Model\CountryTranslation';
 }