+++ changelog.md
@@ -4,7 +4,7 @@
 - Add PHP7 type-hints [#557](https://github.com/dimsav/laravel-translatable/pull/557)
 
 ## v10.0.0
-- Add `Dimsav\Translatable\Locales` helper class [#574](https://github.com/dimsav/laravel-translatable/pull/574)
+- Add `Etchfoda\Translatable\Locales` helper class [#574](https://github.com/dimsav/laravel-translatable/pull/574)
 - Fix `getRelationKey()` [#575](https://github.com/dimsav/laravel-translatable/pull/575)
 - drop custom save method in favor of saved event listener [#567](https://github.com/dimsav/laravel-translatable/pull/575)